JS: Move some TypeScript classes into .ts.extractor
